Fixed tests being placed under lib. Moved to src
