+++ package.json
@@ -18,6 +18,7 @@
   },
   "devDependencies": {
     "babel-eslint": "^3.1.9",
+    "babel-jest": "^5.2.0",
     "eslint": "^0.21.2",
     "eslint-plugin-react": "^2.3.0",
     "gulp": "^3.8.11",
@@ -39,6 +40,7 @@
     "watch": "gulp watch:lib"
   },
   "jest": {
+    "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
     "unmockedModulePathPatterns": [
       "<rootDir>/node_modules/react"
     ]
