+++ src/ContextExample.js
@@ -1,4 +1,5 @@
 var React = require('react');
+var One = require('./One');
 
 var ContextExample = React.createClass({
 	propTypes: {
@@ -21,71 +22,6 @@ var ContextExample = React.createClass({
 		);
 	}
 	
-});
-
-var One = React.createClass({
-	propTypes: {
-		b: React.PropTypes.string
-	},
-	contextTypes: {
-		a: React.PropTypes.string
-	},
-	childContextTypes: {
-		b: React.PropTypes.string
-	},
-	getChildContext() {
-		return {
-			b: this.props.b
-		}
-	},
-	render() {
-		return (
-			<div>
-				One
-				({this.context.a})
-				<Two a="Zed" />
-			</div>
-		);
-	}
-});
-
-var Two = React.createClass({
-	contextTypes: {
-		a: React.PropTypes.string,
-		b: React.PropTypes.string
-	},
-	childContextTypes: {
-		a: React.PropTypes.string
-	},
-	getChildContext() {
-		return {
-			a: this.props.a || this.context.a
-		}
-	},
-	render() {
-		return (
-			<div>
-				Two
-				({this.context.a}, {this.context.b})
-				<Three />
-			</div>
-		);
-	}
-});
-
-var Three = React.createClass({
-	contextTypes: {
-		a: React.PropTypes.string,
-		b: React.PropTypes.string
-	},
-	render() {
-		return (
-			<div>
-				Three
-				({this.context.a}, {this.context.b})
-			</div>
-		);
-	}
 });
 
 export default ContextExample;
